refactor(server): move healthcheck
